+++ src/MTProtoTools/UpdateHandler.php
@@ -67,6 +67,7 @@ use danog\MadelineProto\EventHandler\Message\Service\DialogTopicCreated;
 use danog\MadelineProto\EventHandler\Message\Service\DialogTopicEdited;
 use danog\MadelineProto\EventHandler\Message\Service\DialogUserTyping;
 use danog\MadelineProto\EventHandler\Message\Service\DialogWebView;
+use danog\MadelineProto\EventHandler\Message\Service\SecretChat\SecretUserTyping;
 use danog\MadelineProto\EventHandler\Privacy;
 use danog\MadelineProto\EventHandler\Query\ChatButtonQuery;
 use danog\MadelineProto\EventHandler\Query\ChatGameQuery;
@@ -677,11 +678,10 @@ trait UpdateHandler
                     $message,
                     $info,
                 ),
-                'decryptedMessageActionTyping' => new DialogUserTyping(
+                'decryptedMessageActionTyping' => new SecretUserTyping(
                     $this,
                     $message,
                     $info,
-                    Action::fromRawAction($action['action'])
                 ),
                 default => null
             };
